AudioTabController: add force device on microphone
This is a fix for VRChat's mic switch issue. When it's figured out on their end, this could be removed. It was found that disabling the "launch" mic would change the device used by VRChat, so this creates a "only-one-left" situation with the current device.

I'd like this tested in various situations because it may cause issues in other apps like VOIP.
 
- Added a button to trigger forceCurrentRecordingDevice,
- forceCurrentRecordingDevice loops through devices, except for the current, and disables them with a 30ms delay between.
- Added documentation in README.md.